pyproject.toml
[project]
name = "dataset-mcp"
version = "0.1.0"
description = "知识库目录树生成工具"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.9",
    "fastmcp>=2.6.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.27.0",
    "jieba>=0.42.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dataset"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.black]
line-length = 88
target-version = ['py312']
